+++ src/compliance_tests/TestIso_8_4_2.cpp
@@ -80,7 +80,7 @@ class TestIso_8_4_2 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 2; i++)
+            for (size_t i = 0; i < 2; i++)
             {
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
@@ -150,7 +150,7 @@ class TestIso_8_4_2 : public test::TestBase
             }
 
             Bit *last_delim_bit = drv_bit_frm->GetBitOf(7, BitKind::ErrDelim);
-            int last_delim_index = drv_bit_frm->GetBitIndex(last_delim_bit);
+            size_t last_delim_index = drv_bit_frm->GetBitIndex(last_delim_bit);
             drv_bit_frm->FlipBitAndCompensate(last_delim_bit, dut_input_delay);
 
             mon_bit_frm->InsertOvrlFrm(last_delim_index + 1);
