+++ src/compliance_tests/TestIso_8_4_4.cpp
@@ -83,7 +83,7 @@ class TestIso_8_4_4 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 3; i++)
+            for (size_t i = 0; i < 3; i++)
             {
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
@@ -138,7 +138,7 @@ class TestIso_8_4_4 : public test::TestBase
                 bit_to_corrupt = drv_bit_frm->GetBitOf(5, BitKind::OvrlFlag);
             }
 
-            int bit_index = drv_bit_frm->GetBitIndex(bit_to_corrupt);
+            size_t bit_index = drv_bit_frm->GetBitIndex(bit_to_corrupt);
             bit_to_corrupt->val_ = BitVal::Recessive;
 
             drv_bit_frm->InsertActErrFrm(bit_index + 1);
