+++ src/compliance_tests/TestIso_8_7_4.cpp
@@ -81,14 +81,14 @@ class TestIso_8_7_4 : public test::TestBase
         {
             FillTestVariants(VariantMatchType::Common);
 
-            int num_elem_tests = nbt.sjw_;
+            size_t num_elem_tests = nbt.sjw_;
             if (nbt.ph2_ < nbt.sjw_)
                 num_elem_tests = nbt.ph2_;
 
-            for (int i = 0; i < num_elem_tests; i++)
+            for (size_t i = 0; i < num_elem_tests; i++)
             {
                 ElemTest test = ElemTest(i + 1);
-                test.e_ = i + 1;
+                test.e_ = static_cast<int>(i + 1);
                 AddElemTest(TestVariant::Common, std::move(test));
             }
 
@@ -127,7 +127,7 @@ class TestIso_8_7_4 : public test::TestBase
              *************************************************************************************/
             Bit *bit_to_shorten;
             Bit *next_bit;
-            int bit_index;
+            size_t bit_index;
             do
             {
                 bit_to_shorten = drv_bit_frm->GetRandBitOf(BitKind::BaseIdent);
@@ -136,9 +136,9 @@ class TestIso_8_7_4 : public test::TestBase
             } while (!(bit_to_shorten->val_ == BitVal::Recessive &&
                         next_bit->val_ == BitVal::Dominant));
 
-            bit_to_shorten->ShortenPhase(BitPhase::Ph2, elem_test.e_);
+            bit_to_shorten->ShortenPhase(BitPhase::Ph2, static_cast<size_t>(elem_test.e_));
 
-            mon_bit_frm->GetBit(bit_index)->ShortenPhase(BitPhase::Ph2, elem_test.e_ - 1);
+            mon_bit_frm->GetBit(bit_index)->ShortenPhase(BitPhase::Ph2, static_cast<size_t>(elem_test.e_ - 1));
             mon_bit_frm->GetBit(bit_index + 1)->ShortenPhase(BitPhase::Sync, 1);
 
             drv_bit_frm->GetBitOf(0, BitKind::Ack)->val_ = BitVal::Dominant;
